+++ blocks/list/src/index.jsx
@@ -1,0 +1,6 @@
+import { bootstrap } from '@appsemble/react';
+import { provideIntl } from '@appsemble/react/intl';
+
+import ListBlock from './components/ListBlock';
+
+bootstrap(provideIntl(ListBlock));