+++ packages/editor/actions/index.js
@@ -1,4 +1,5 @@
 export { default as apps } from './apps';
 export { default as db } from './db';
-export { default as user } from './user';
 export { default as message } from './message';
+export { default as openApi } from './openApi';
+export { default as user } from './user';