+++ packages/server/api/components/requestBodies/blockDefinition.js
@@ -1,0 +1,10 @@
+export default {
+  description: 'A block definition',
+  content: {
+    'application/json': {
+      schema: {
+        $ref: '#/components/schemas/BlockDefinition',
+      },
+    },
+  },
+};