+++ packages/server/api/components/responses/default.js
@@ -1,0 +1,10 @@
+export default {
+  description: 'The default error response.',
+  content: {
+    'application/json': {
+      schema: {
+        $ref: '#/components/schemas/Error',
+      },
+    },
+  },
+};