+++ packages/server/api/components/schemas/Resource.js
@@ -1,0 +1,10 @@
+export default {
+  type: 'object',
+  additionalProperties: true,
+  properties: {
+    id: {
+      type: 'number',
+      readOnly: true,
+    },
+  },
+};