+++ packages/server/api/components/schemas/index.js
@@ -1,0 +1,23 @@
+import App from './App';
+import BlockDefinition from './BlockDefinition';
+import Block from './Block';
+import BlockVersion from './BlockVersion';
+import Color from './Color';
+import Error from './Error';
+import Page from './Page';
+import Resource from './Resource';
+import ResourceDefinition from './ResourceDefinition';
+import Theme from './Theme';
+
+export default {
+  App,
+  BlockDefinition,
+  Block,
+  BlockVersion,
+  Color,
+  Error,
+  Page,
+  Resource,
+  ResourceDefinition,
+  Theme,
+};