+++ packages/server/api/tags/app.js
@@ -1,0 +1,4 @@
+export default {
+  name: 'app',
+  description: 'Operations related to the App definition',
+};